requirements.txt
tk
pynput
mss
Pillow
openai
python-dotenv
ollama
